+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Added
 
 - PNPM v9 lockfile support
+- Support for parsing `go.mod` files with a Go directive of version 1.17 and higher
 
 ## Changed